+++ src/NativeProviders/OpenBLAS/capabilities.cpp
@@ -1,5 +1,8 @@
 #include "wrapper_common.h"
-#include "cblas.h"
+#include "blas.h"
+#if defined(BLIS_INT_TYPE_SIZE)
+#include "blis.h"
+#endif
 
 #ifdef __cplusplus
 extern "C" {
@@ -55,22 +58,40 @@ extern "C" {
 
 	DLLEXPORT void set_max_threads(const blasint num_threads)
 	{
+#if defined(BLIS_INT_TYPE_SIZE)
+		bli_thread_set_num_threads((dim_t)num_threads);
+#else
 		openblas_set_num_threads(num_threads);
+#endif
 	}
 
 	DLLEXPORT char* get_build_config()
 	{
+#if defined(BLIS_INT_TYPE_SIZE)
+		static char config[] = "AOCL BLIS";
+		return config;
+#else
 		return openblas_get_config();
+#endif
 	}
 
 	DLLEXPORT char* get_cpu_core()
 	{
+#if defined(BLIS_INT_TYPE_SIZE)
+		static char core[] = "AOCL";
+		return core;
+#else
 		return openblas_get_corename();
+#endif
 	}
 
 	DLLEXPORT int get_parallel_type()
 	{
+#if defined(BLIS_INT_TYPE_SIZE)
+		return 0;
+#else
 		return openblas_get_parallel();
+#endif
 	}
 
 #ifdef __cplusplus